pyproject.toml
[project]
name = "ai-shopping-cart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "openai==1.97.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
